+++ CHANGELOG.md
@@ -208,3 +208,4 @@
 * Define `Weighted` type family and define instance for `Graph`
 * Add `Leap` to `Edges` to indicate a connection with only the first vertex
 * Remove `Graph` structure and `Weighted` type family
+* Define `Reverse` type to run `Traversable` in reverse order with `Backwards`